Changed return type of method CalculateMin

+++ Cwiczenia2/StatisticsHelper.cs
@@ -24,7 +24,7 @@ public class StatisticsHelper
         return 67;
     }
 
-    public int CalculateMin(int[] values)
+    public double CalculateMin(int[] values)
     {
         return 6;
     }
